Make heartbeathandler shared struct
