+++ src/gateway/default/heartbeat.rs
@@ -1,22 +1,7 @@
-use crate::types;
-
 use super::*;
 
-/// Handles sending heartbeats to the gateway in another thread
-#[allow(dead_code)] // FIXME: Remove this, once HeartbeatHandler is used
-#[derive(Debug)]
-pub struct DefaultHeartbeatHandler {
-    /// How ofter heartbeats need to be sent at a minimum
-    pub heartbeat_interval: Duration,
-    /// The send channel for the heartbeat thread
-    pub send: Sender<HeartbeatThreadCommunication>,
-    /// The handle of the thread
-    handle: JoinHandle<()>,
-}
-
-impl HeartbeatHandlerCapable<WebSocketStream<MaybeTlsStream<TcpStream>>>
-    for DefaultHeartbeatHandler
-{
+#[async_trait]
+impl HeartbeatHandlerCapable<WebSocketStream<MaybeTlsStream<TcpStream>>> for HeartbeatHandler {
     fn new(
         heartbeat_interval: Duration,
         websocket_tx: Arc<
@@ -28,12 +13,12 @@ impl HeartbeatHandlerCapable<WebSocketStream<MaybeTlsStream<TcpStream>>>
             >,
         >,
         kill_rc: tokio::sync::broadcast::Receiver<()>,
-    ) -> DefaultHeartbeatHandler {
+    ) -> HeartbeatHandler {
         let (send, receive) = tokio::sync::mpsc::channel(32);
         let kill_receive = kill_rc.resubscribe();
 
         let handle: JoinHandle<()> = task::spawn(async move {
-            DefaultHeartbeatHandler::heartbeat_task(
+            HeartbeatHandler::heartbeat_task(
                 websocket_tx,
                 heartbeat_interval,
                 receive,
@@ -55,94 +40,5 @@ impl HeartbeatHandlerCapable<WebSocketStream<MaybeTlsStream<TcpStream>>>
 
     fn get_heartbeat_interval(&self) -> Duration {
         self.heartbeat_interval
-    }
-}
-
-impl DefaultHeartbeatHandler {
-    /// The main heartbeat task;
-    ///
-    /// Can be killed by the kill broadcast;
-    /// If the websocket is closed, will die out next time it tries to send a heartbeat;
-    pub async fn heartbeat_task(
-        websocket_tx: Arc<
-            Mutex<
-                SplitSink<
-                    WebSocketStream<MaybeTlsStream<TcpStream>>,
-                    tokio_tungstenite::tungstenite::Message,
-                >,
-            >,
-        >,
-        heartbeat_interval: Duration,
-        mut receive: tokio::sync::mpsc::Receiver<HeartbeatThreadCommunication>,
-        mut kill_receive: tokio::sync::broadcast::Receiver<()>,
-    ) {
-        let mut last_heartbeat_timestamp: Instant = time::Instant::now();
-        let mut last_heartbeat_acknowledged = true;
-        let mut last_seq_number: Option<u64> = None;
-
-        loop {
-            if kill_receive.try_recv().is_ok() {
-                trace!("GW: Closing heartbeat task");
-                break;
-            }
-
-            let timeout = if last_heartbeat_acknowledged {
-                heartbeat_interval
-            } else {
-                // If the server hasn't acknowledged our heartbeat we should resend it
-                Duration::from_millis(HEARTBEAT_ACK_TIMEOUT)
-            };
-
-            let mut should_send = false;
-
-            tokio::select! {
-                () = sleep_until(last_heartbeat_timestamp + timeout) => {
-                    should_send = true;
-                }
-                Some(communication) = receive.recv() => {
-                    // If we received a seq number update, use that as the last seq number
-                    if communication.sequence_number.is_some() {
-                        last_seq_number = communication.sequence_number;
-                    }
-
-                    if let Some(op_code) = communication.op_code {
-                        match op_code {
-                            GATEWAY_HEARTBEAT => {
-                                // As per the api docs, if the server sends us a Heartbeat, that means we need to respond with a heartbeat immediately
-                                should_send = true;
-                            }
-                            GATEWAY_HEARTBEAT_ACK => {
-                                // The server received our heartbeat
-                                last_heartbeat_acknowledged = true;
-                            }
-                            _ => {}
-                        }
-                    }
-                }
-            }
-
-            if should_send {
-                trace!("GW: Sending Heartbeat..");
-
-                let heartbeat = types::GatewayHeartbeat {
-                    op: GATEWAY_HEARTBEAT,
-                    d: last_seq_number,
-                };
-
-                let heartbeat_json = serde_json::to_string(&heartbeat).unwrap();
-
-                let msg = tokio_tungstenite::tungstenite::Message::text(heartbeat_json);
-
-                let send_result = websocket_tx.lock().await.send(msg).await;
-                if send_result.is_err() {
-                    // We couldn't send, the websocket is broken
-                    warn!("GW: Couldnt send heartbeat, websocket seems broken");
-                    break;
-                }
-
-                last_heartbeat_timestamp = time::Instant::now();
-                last_heartbeat_acknowledged = false;
-            }
-        }
     }
 }
